requirements.txt
# Core dependencies
requests>=2.31.0
numpy>=1.24.0

# Video processing
yt-dlp>=2023.0.0

# Optional: For advanced audio analysis
librosa>=0.10.0
soundfile>=0.12.0

# Optional: For AI-powered highlight detection
openai>=1.0.0  # For GPT-4 Vision analysis
anthropic>=0.25.0  # For Claude vision analysis

# Optional: For speech recognition (future subtitle feature)
openai-whisper>=20230918

# Web interface
flask>=3.0.0